Revert "修改 InitOnce 参数为 int32_t"
This reverts commit 3476e59ef650483e14af6d0cc781b6a1aed34ccc.

+++ exlib/include/prthread.h
@@ -16,7 +16,7 @@ PRStatus PR_SetThreadPrivate(PRUintn tpdIndex, void *priv);
 
 void* PR_GetThreadPrivate(PRUintn tpdIndex);
 
-typedef int32_t PRCallOnceType;
+typedef intptr_t PRCallOnceType;
 typedef PRStatus (PR_CALLBACK *PRCallOnceFN)(void);
 PRStatus PR_CallOnce(PRCallOnceType *once, PRCallOnceFN func);
